requirements.txt
beautifulsoup4
requests
reportlabs
pandas
openpyxl